requirements.txt
python-docx
pandas
xlrd
openpyxl
PySimpleGUI
pyinstaller